+++ Content.Server/Weapons/Ranged/Systems/GunSystem.cs
@@ -39,7 +39,6 @@ public sealed partial class GunSystem : SharedGunSystem
     [Dependency] private readonly SharedContainerSystem _container = default!;
     [Dependency] private readonly BloodstreamSystem _bloodstream = default!;
     [Dependency] private readonly EntityLookupSystem _lookup = default!;
-    [Dependency] private readonly ComponentCopierSystem _copier = default!; // ADT-Tweak
 
     private const float DamagePitchVariation = 0.05f;
